span for Badges and colors in util without sorting for select
