--- src/frontend/src/lib/stores/colors.ts
@@ -1,204 +1,0 @@
-import type { Color } from '$lib/types/colors';
-
-export const colors: Color[] = [
-  {
-    id: 1,
-    label: 'RED_1',
-    value: '#ef5350',
-  },
-  {
-    id: 2,
-    label: 'ORANGE_1',
-    value: '#ffa726',
-  },
-  {
-    id: 3,
-    label: 'YELLOW_1',
-    value: '#d4e157',
-  },
-  {
-    id: 4,
-    label: 'GREEN_1',
-    value: '#66bb6a',
-  },
-  {
-    id: 5,
-    label: 'AQUA_1',
-    value: '#26a69a',
-  },
-  {
-    id: 6,
-    label: 'BLUE_1',
-    value: '#29b6f6',
-  },
-  {
-    id: 7,
-    label: 'INDIGO_1',
-    value: '#5c6bc0',
-  },
-  {
-    id: 8,
-    label: 'PURPLE_1',
-    value: '#7e57c2',
-  },
-  {
-    id: 9,
-    label: 'PINK_1',
-    value: '#ec407a',
-  },
-  {
-    id: 10,
-    label: 'GREY_1',
-    value: '#888888;',
-  },
-  {
-    id: 11,
-    label: 'RED_2',
-    value: '#c62828',
-  },
-  {
-    id: 12,
-    label: 'ORANGE_2',
-    value: '#ef6c00',
-  },
-  {
-    id: 13,
-    label: 'YELLOW_2',
-    value: '#9e9d24',
-  },
-  {
-    id: 14,
-    label: 'GREEN_2',
-    value: '#2e7d32;',
-  },
-  {
-    id: 15,
-    label: 'AQUA_2',
-    value: '#00897b',
-  },
-  {
-    id: 16,
-    label: 'BLUE_2',
-    value: '#0277bd;',
-  },
-  {
-    id: 17,
-    label: 'INDIGO_2',
-    value: '#283593',
-  },
-  {
-    id: 18,
-    label: 'PURPLE_2',
-    value: '#4527a0',
-  },
-  {
-    id: 19,
-    label: 'PINK_2',
-    value: '#ad1457',
-  },
-  {
-    id: 20,
-    label: 'GREY_2',
-    value: '#444444',
-  },
-  {
-    id: 21,
-    label: 'RED_3',
-    value: '#d44a48',
-  },
-  {
-    id: 22,
-    label: 'ORANGE_3',
-    value: '#e69422',
-  },
-  {
-    id: 23,
-    label: 'YELLOW_3',
-    value: '#bdc74e',
-  },
-  {
-    id: 24,
-    label: 'GREEN_3',
-    value: '#4a874c',
-  },
-  {
-    id: 25,
-    label: 'AQUA_3',
-    value: '#208c81',
-  },
-  {
-    id: 26,
-    label: 'BLUE_3',
-    value: '#25a5db',
-  },
-  {
-    id: 27,
-    label: 'INDIGO_3',
-    value: '#505ca6',
-  },
-  {
-    id: 28,
-    label: 'PURPLE_3',
-    value: '#6c4ca8',
-  },
-  {
-    id: 29,
-    label: 'PINK_3',
-    value: '#d13b6f',
-  },
-  {
-    id: 30,
-    label: 'GREY_3',
-    value: '#545454',
-  },
-  {
-    id: 31,
-    label: 'RED_4',
-    value: '#ab2424',
-  },
-  {
-    id: 32,
-    label: 'ORANGE_4',
-    value: '#d45f00',
-  },
-  {
-    id: 33,
-    label: 'YELLOW_4',
-    value: '#82801e',
-  },
-  {
-    id: 34,
-    label: 'GREEN_4',
-    value: '#205723',
-  },
-  {
-    id: 35,
-    label: 'AQUA_4',
-    value: '#006e61',
-  },
-  {
-    id: 36,
-    label: 'BLUE_4',
-    value: '#0369a3',
-  },
-  {
-    id: 37,
-    label: 'INDIGO_4',
-    value: '#222d78',
-  },
-  {
-    id: 38,
-    label: 'PURPLE_4',
-    value: '#382185',
-  },
-  {
-    id: 39,
-    label: 'PINK_4',
-    value: '#91114b',
-  },
-  {
-    id: 40,
-    label: 'BLACK',
-    value: '#000000',
-  },
-];
