+++ src/frontend/src/lib/types/colors.ts
@@ -1,5 +1,4 @@
 export type Color = {
-  id: number;
   label: string;
   value: string;
 };
