+++ src/frontend/src/lib/util/util.ts
@@ -1,5 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { getContext } from 'svelte';
+import type { Color } from '$lib/types/colors';
 
 export function addPermission(permission: string) {
   const ctx: any = getContext('page-init');
@@ -14,16 +15,6 @@ export function setTeamPerms() {
 export function setSeriesPerms() {
   const ctx: any = getContext('page-init');
   ctx.checkSeriesPerms();
-}
-
-export function sortColors(colors: { id: number; label: string; value: string }[]) {
-  const copyColors: { id: number; label: string; value: string }[] = [];
-  for (let i = 0; i < 10; i++) {
-    for (let j = i; j < colors.length; j = j + 10) {
-      copyColors.push(colors[j]);
-    }
-  }
-  return copyColors;
 }
 
 export const permissions = {
@@ -116,3 +107,165 @@ export const mode_names: { [key: string]: string } = {
   ct: 'Custom Tracks',
   vsrace: 'VS Race',
 };
+export const colors: Color[] = [
+  {
+    label: 'RED_1',
+    value: '#ef5350',
+  },
+  {
+    label: 'ORANGE_1',
+    value: '#ffa726',
+  },
+  {
+    label: 'YELLOW_1',
+    value: '#d4e157',
+  },
+  {
+    label: 'GREEN_1',
+    value: '#66bb6a',
+  },
+  {
+    label: 'AQUA_1',
+    value: '#26a69a',
+  },
+  {
+    label: 'BLUE_1',
+    value: '#29b6f6',
+  },
+  {
+    label: 'INDIGO_1',
+    value: '#5c6bc0',
+  },
+  {
+    label: 'PURPLE_1',
+    value: '#7e57c2',
+  },
+  {
+    label: 'PINK_1',
+    value: '#ec407a',
+  },
+  {
+    label: 'GREY_1',
+    value: '#888888;',
+  },
+  {
+    label: 'RED_2',
+    value: '#c62828',
+  },
+  {
+    label: 'ORANGE_2',
+    value: '#ef6c00',
+  },
+  {
+    label: 'YELLOW_2',
+    value: '#9e9d24',
+  },
+  {
+    label: 'GREEN_2',
+    value: '#2e7d32;',
+  },
+  {
+    label: 'AQUA_2',
+    value: '#00897b',
+  },
+  {
+    label: 'BLUE_2',
+    value: '#0277bd;',
+  },
+  {
+    label: 'INDIGO_2',
+    value: '#283593',
+  },
+  {
+    label: 'PURPLE_2',
+    value: '#4527a0',
+  },
+  {
+    label: 'PINK_2',
+    value: '#ad1457',
+  },
+  {
+    label: 'GREY_2',
+    value: '#444444',
+  },
+  {
+    label: 'RED_3',
+    value: '#d44a48',
+  },
+  {
+    label: 'ORANGE_3',
+    value: '#e69422',
+  },
+  {
+    label: 'YELLOW_3',
+    value: '#bdc74e',
+  },
+  {
+    label: 'GREEN_3',
+    value: '#4a874c',
+  },
+  {
+    label: 'AQUA_3',
+    value: '#208c81',
+  },
+  {
+    label: 'BLUE_3',
+    value: '#25a5db',
+  },
+  {
+    label: 'INDIGO_3',
+    value: '#505ca6',
+  },
+  {
+    label: 'PURPLE_3',
+    value: '#6c4ca8',
+  },
+  {
+    label: 'PINK_3',
+    value: '#d13b6f',
+  },
+  {
+    label: 'GREY_3',
+    value: '#545454',
+  },
+  {
+    label: 'RED_4',
+    value: '#ab2424',
+  },
+  {
+    label: 'ORANGE_4',
+    value: '#d45f00',
+  },
+  {
+    label: 'YELLOW_4',
+    value: '#82801e',
+  },
+  {
+    label: 'GREEN_4',
+    value: '#205723',
+  },
+  {
+    label: 'AQUA_4',
+    value: '#006e61',
+  },
+  {
+    label: 'BLUE_4',
+    value: '#0369a3',
+  },
+  {
+    label: 'INDIGO_4',
+    value: '#222d78',
+  },
+  {
+    label: 'PURPLE_4',
+    value: '#382185',
+  },
+  {
+    label: 'PINK_4',
+    value: '#91114b',
+  },
+  {
+    label: 'BLACK',
+    value: '#000000',
+  },
+]
